move callbacks.py to _internal root dir
