+++ nanome/_internal/callbacks.py
@@ -1,7 +1,6 @@
 
 
 import nanome
-from nanome.util.stream import StreamCreationError
 from .enums import Hashes
 import logging
 
@@ -28,17 +27,16 @@ def connect(network, arg, request_id):
 
 
 def receive_create_stream_result(network, result, request_id):
-    if result[0] != StreamCreationError.NoError:
-        network._call(request_id, None, result[0])
+    
+    network._call(request_id, None, result[0])
 
-        if result[0] == StreamCreationError.UnsupportedStream:
-            logger.error("Tried to create an unsupported type of stream")
+        
+        logger.error("Tried to create an unsupported type of stream")
         return
 
     from nanome.api.streams import Stream
     stream = Stream(network, result[1], result[2], result[3])
-    network._call(request_id, stream, StreamCreationError.NoError)
-
+    
 
 def feed_stream(network, result, request_id):
     from nanome.api.streams import Stream
