+++ nanome/_internal/plugin.py
@@ -1,7 +1,7 @@
 from nanome._internal import network as Network
 from nanome._internal.process import ProcessManager
 from nanome._internal.network import PluginNetwork
-from nanome._internal.network.enums import Hashes
+from nanome._internal.enums import Hashes
 from nanome._internal.serializers import CommandMessageSerializer
 from nanome._internal.util.type_serializers import TypeSerializer
 from nanome._internal.logs import LogsManager
