+++ nanome/_internal/ui/models.py
@@ -72,13 +72,12 @@ class _Button(_UIBase):
         self._pressed_callback = func
 
     def _register_hover_callback(self, func):
-        import nanome
-        message_callbacks = nanome._internal.network.enums.Messages
-        message_serializers = nanome._internal.message_serializers
+        from nanome._internal import enums, message_serializers, network
+        message_callbacks = enums.Messages
         if func == None and self._hover_callback == None:  # Low hanging filter but there may be others
             return
         try:
-            nanome._internal.network.PluginNetwork._instance._send(
+            network.PluginNetwork._instance._send(
                 message_callbacks.hook_ui_callback,
                 (message_serializers.UIHook.Type.button_hover, self._content_id),
                 False)
